Knight Changes
Previously, the knight could not do its combo action of moving and being able to capture a piece in one action, if it was the final action since the turn would then end after the initial movement. If it is the final turn, the user will have an option to make another capture. This knight combo move also does not work on diagonal pieces, so took away that functionality. I had to make it so the knight can move diagonally to capture if it is not the last move(so this would count as a separate action). If this is the last action avaiable, the only possible moves(captures) are the adjacent pieces.